Attempt to circumvent puppeteer check
