+++ src/test/url.ts
@@ -1,0 +1,3 @@
+const url = (path: string) => `http://localhost:${puppeteerConfig.server.port}/${path.replace(/^\/+/, '')}`;
+
+export default url;
